(Client+Server) Fixed occasional desync that caused game aborts - removed debugging code
This occurred when the user managed to place a brick while previous brick placements were being sent to the server. Obviously, this occured much more likely when the server was slow or when the users internet connection was bad / slow.
It was fixed by correcting how these brick placements are handled in TransactionQueue.